pyproject.toml
[tool.poetry]
name = "translator-404-bot"
version = "1.1.1-dev"
description = "Telegram bot for translating messages in chats with text in another language (404)"
authors = ["lexxai <lexxai@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.39.0"
googletrans = "^4.0.2"
langdetect = "^1.0.9"
python-dotenv = "^1.0.1"
cryptg = "^0.5.0.post0"
pillow = "^11.1.0"
aiohttp = "^3.11.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
